WebEngine-specific: don't cache MDX video in a temporary directory
gdvideo:// links are playable in the Qt WebEngine version, but not in
the Qt WebKit version.

ArticleResourceReply is a sequential device and thus does not support
seeking. Qt WebEngine or Chromium internals attempt and fail to seek in
the video device unless instructed to preload the entire video.
Force the preloading to prevent errors and allow playing the video.
The rest of the article is loaded and displayed before the video is
preloaded, therefore this commit still improves performance compared to
previous blocking of article loading while getCachedFileName() saves the
video file in a temporary directory.

The video data is stored inside a dictionary file on disk, which is
subject to operating system page cache. Therefore, eliminating the
intermediate cache in a temporary directory shouldn't slow down repeated
loading of large videos.

b10cdf685c8c02d2be2f1ae4f1e03a4e02db48d8 introduced the
getCachedFileName() workaround and made a widely used in GoldenDict code
class Mutex recursive. The primary motivation for this commit is that
disabling the workaround allows to make Mutex nonrecursive again in the
Qt WebEngine version. This improves overall GoldenDict performance by
eliminating the recursiveness overhead. It also eliminates
QMutex::Recursive deprecation warnings and paves the way to Qt 6
porting. Disable the warning in the Qt WebKit version via a GCC pragma,
because Qt WebKit is incompatible with Qt 6, where the deprecated
feature is no longer available.